rn,tracks: do not load stream effects on mobile
