+++ react/features/base/tracks/loadEffects.native.js
@@ -1,0 +1,11 @@
+// @flow
+
+/**
+ * Loads the enabled stream effects.
+ *
+ * @param {Object} store - The Redux store.
+ * @returns {Promsie} - A Promise which resolves when all effects are created.
+ */
+export default function loadEffects(store: Object): Promise<any> { // eslint-disable-line no-unused-vars
+    return Promise.resolve();
+}
